Refactor API, enhance performance, and add new utilities
- Refactor API to improve consistency and simplify function calls
- Enhance performance by converting all types to bytes
- Add new API functions for more functionality
- Update documentation to reflect API changes and new features
- Fix minor bugs and improve error handling

+++ safe-vk/src/responses/message.rs
@@ -30,7 +30,7 @@ pub struct ReplyMessage<T> {
     /// Time the message was sent in Unixtime.
     pub date: i64,
     /// Destination identifier.
-    pub peer_id: i32,
+    pub peer_id: i64,
     /// Sender identifier.
     pub from_id: i32,
     /// Message text.
@@ -48,7 +48,7 @@ pub struct PersonalMessage<T> {
     /// Time the message was sent in Unixtime.
     pub date: i64,
     /// Destination identifier.
-    pub peer_id: i32,
+    pub peer_id: i64,
     /// Sender identifier.
     pub from_id: i32,
     /// Message text.
